KVM: Change return type of kvm_arch_vm_ioctl() to "int"
JIRA: https://issues.redhat.com/browse/RHEL-14702

All kvm_arch_vm_ioctl() implementations now only deal with "int"
types as return values, so we can change the return type of these
functions to use "int" instead of "long".

Acked-by: Anup Patel <anup@brainfault.org>
Message-Id: <20230208140105.655814-7-thuth@redhat.com>
(cherry picked from commit d8708b80fa0e6e21bc0c9e7276ad0bccef73b6e7)

[RHEL-only: adjacent functions already added downstream, no RISC-V]

+++ arch/arm64/kvm/arm.c
@@ -1571,8 +1571,7 @@ static int kvm_vm_set_attr(struct kvm *kvm, struct kvm_device_attr *attr)
 	}
 }
 
-long kvm_arch_vm_ioctl(struct file *filp,
-		       unsigned int ioctl, unsigned long arg)
+int kvm_arch_vm_ioctl(struct file *filp, unsigned int ioctl, unsigned long arg)
 {
 	struct kvm *kvm = filp->private_data;
 	void __user *argp = (void __user *)arg;
